+++ src/main/java/wooteco/subway/admin/controller/LineStationController.java
@@ -4,7 +4,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import wooteco.subway.admin.dto.controller.request.LineStationCreateControllerRequest;
 import wooteco.subway.admin.dto.controller.response.LineStationCreateControllerResponse;
-import wooteco.subway.admin.dto.response.LineStationCreateServiceResponse;
+import wooteco.subway.admin.dto.service.response.LineStationCreateServiceResponse;
 import wooteco.subway.admin.dto.service.response.StationCreateServiceResponse;
 import wooteco.subway.admin.dto.view.request.LineStationCreateViewRequest;
 import wooteco.subway.admin.service.LineService;
